Revert "Yashpal branch (#2)"
This reverts commit b1a19055447fad71816c07e08587b317f3224d49.
