+++ client/src/index.js
@@ -7,15 +7,14 @@ import {
   useRouteError,
 } from "react-router-dom";
 import { Container, Typography } from "@mui/material";
-import Login from "./Components/Auth/Login";
-import Signup from "./Components/Auth/Signup";
-import ForgotPassword from "./Components/Auth/ForgotPassword";
-import Home from "./Components/Home/Home";
+import Login from "./Components/Login";
+import Signup from "./Components/Signup";
+import ForgotPassword from "./Components/ForgotPassword";
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <Navigate to={"/home"} />,
+    element: <Navigate to={"/login"} />,
     errorElement: <ErrorBoundry />,
   },
   {
@@ -32,7 +31,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/home",
-    element: <Home />,
+    element: <h2 className="text-warning">Welcome Home</h2>,
   },
 ]);
