+++ models/user.schema.js
@@ -13,15 +13,7 @@ const userSchema = new mongoose.Schema({
   },
   password: {
     type: String,
-    required: function () {
-      return !this.googleAuth; // required if google auth not provided
-    },
-  },
-  googleAuth: {
-    type: Boolean,
-    required: function () {
-      return !this.password; // required if password not provided
-    },
+    required: true,
   },
   lastUpdated: {
     type: Date,
