+++ routes/user.route.js
@@ -1,19 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const {
-  setUsername,
-  signInWithGoogle,
-  Signup,
-  Login,
-  ChangePassword,
-  checkSession,
-} = require("../controllers/user.controller");
+const { Signup, Login, ChangePassword } = require("../controllers/user.controller");
 
-router.post("/googleAuth", signInWithGoogle);
-router.post("/setUsername", setUsername);
 router.post("/signup", Signup);
 router.post("/login", Login);
 router.post("/change-password", ChangePassword);
-router.post("/checkSession", checkSession);
 
 module.exports = router;
